fix(supplier): fix broken imports and handle invalid ids

+++ src/modules/supplier/repo.js
@@ -1,4 +1,6 @@
-import Supplier from "./model";
+import Supplier from "./model.js";
+import { Types } from "mongoose";
+import { BadRequestError } from "../../utils/errors.js";
 class SupplierRepo {
     constructor(model) { this.model = model; }
 
@@ -7,18 +9,30 @@ class SupplierRepo {
     }
 
     async findAll(userId = "") {
+        if (userId && !Types.ObjectId.isValid(userId)) {
+            throw new BadRequestError("Invalid user ID");
+        }
         return await this.model.find({ userId });
     }
 
     async findById(id) {    
+        if (!Types.ObjectId.isValid(id)) {
+            throw new BadRequestError("Invalid supplier ID");
+        }
         return await this.model.findById(id);
     }
 
     async update(id, data) {
+        if (!Types.ObjectId.isValid(id)) {
+            throw new BadRequestError("Invalid supplier ID");
+        }
         return await this.model.findByIdAndUpdate(id, data, { new: true });
     }
 
     async delete(id) {
+        if (!Types.ObjectId.isValid(id)) {
+            throw new BadRequestError("Invalid supplier ID");
+        }
         return await this.model.findByIdAndDelete(id);
     }
